soporte multiples pantallas android celular

+++ app/src/main/java/com/devworms/editorial/mango/fragments/FavoritosFragment.java
@@ -163,8 +163,10 @@ public class FavoritosFragment extends Fragment {
     public void onPause() {
         super.onPause();
 
-        for(ParseObject receta:mListToDelete){
-            receta.deleteInBackground();
+        if(mListToDelete != null) {
+            for (ParseObject receta : mListToDelete) {
+                receta.deleteInBackground();
+            }
         }
     }
 }
